Revert "feat: add models" (#41)

+++ workbench/database/migrations/2024_04_12_064647_create_orders_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
                 'refunded',
                 'cancelled',
             ])->default('on_hold');
-            $table->float('total');
+            $table->integer('total');
             $table->timestamps();
         });
     }
